boiler_plate/rclone_gdrive/pyproject.toml
[project]
name = "rclone-gdrive"
version = "0.1.0"
description = "Google Drive CLI tool with live progress tracking for upload/download operations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
]

[project.scripts]
gdrive = "main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
